Added or changed methods in the following:
   Constants
Changed the Constants class to be static and include whether prizes are with or without replacement.

   DisplayManager
Added displayTopPrizes.

   FileManager
Added logic to loadPrizeFile to help with math for determining the offer amount later.  Also added comments to explain the variables.

   TotalPrizes
Added variables to easily display the top prizes.
Added variables to help with the logic of determining the offer amount.

   MAIN
Added the function calls to run promotion logic.

Added the following classes:

   Constants
Old class was not static.  Changed the class to static for more efficient const variable lookups.

   OfferGenerator
Static class that handles the logic for determining the offers.

   PromotionManager
Static class that handles the running of the promotion.