Changed the following resource files:
   customerFile.csv
Deleted everything to create the default start file for the Unlock the Box promotion.
Added a information line to state columns, and how file is formatted.

   prizeFile.csv
Revised file to make it the default start prizes for the Unlock the Box promotion.
Added a information line to state columns.

Updated the following classes:

  OfferGenerator
Changed the math to reflect the most recent approved math for offers.

  DisplayManager
Added output and logic for an empty previous winners list and return if empty.
Changed how displayPrizesLeft works to get rid of trailing commas.

  FileManager
Changed file loading methods to get information lines and do nothing with them.
Changed file saving methods to output information lines back into the files.
Changed logic in for loop in saveCustomerFile to avoid a bug which would cause the file to save incorrectly.  This avoids having prizes be on separate lines when they should be on the same line.

   PromotionManager
Added logic to make sure the correct prize is added to customer object.

   TotalPrizes
Added logic in findPrize to only return a prize that has not been won.

   Main
Added logic to check if there are no remaining prizes and skip PromotionManager if true.