Changed the following classes:
  Constants
Added methods so that a defaultPrizeFile and defaultcustomerFile could be created if the files were not present upon loading the application.

  FileManager
Added a pause after failure to load files.
Added logic to loadCustomerFile and loadPrizeFile that would create the default files if files were not present.

Created a installation project to help with installs.
Also added an exe that auto unpacks and installs 2015 visual c++ redist so the app will run properly.